Split dlr_internal.h and dlr.c into multiple files (#83)
